Add files via upload
This is my second project by Oasis Infobyte which is a BMI calculator i used Tkinter and made it through Gui to make it more attractive and feel like a real-world project. In this project a user will be asked to enter the details like his/her weight in KG and height in meter and the application will show the body mass index of that person.